Fixed PWM motor code

+++ include/motors.h
@@ -9,7 +9,7 @@
 #define MOTOR_DIRECTION_R 3
 
 void dc_motors_init();
-void dc_motors_move(const uint8_t direction, const uint8_t power);
+void dc_motors_move(const uint8_t direction, const float power);
 void dc_motors_halt();
 
 void dc_motors_move_for_ms(const uint8_t direction, const uint8_t power, const double time);
